chore(pm): Refactor project management error logic
